fix : 모든 오류 수정

+++ BackEnd/src/main/java/com/ottrade/ottrade/domain/member/dto/AuthDto.java
@@ -8,9 +8,9 @@ import lombok.NoArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
 public class AuthDto {
- 
-	// DTO 통합 
-	
+
+    // DTO 통합
+
     @Getter
     @NoArgsConstructor
     public static class SignUpRequest {
@@ -33,10 +33,10 @@ public class AuthDto {
     @Getter
     @NoArgsConstructor
     public static class LoginRequest {
-    	private String phone;
+        private String phone;
         private String password;
     }
-    
+
     @Getter
     @NoArgsConstructor
     public static class ReissueRequest {
@@ -49,11 +49,12 @@ public class AuthDto {
         private String accessToken;
         private String refreshToken;
     }
-    
+
     @Getter
     @Builder
     public static class UserInfoResponse {
-        private Integer id;
+        // 이 부분의 타입을 Long으로 수정했습니다.
+        private Long id;
         private String email;
         private String nickname;
         private String phone;
@@ -61,7 +62,7 @@ public class AuthDto {
 
         public static UserInfoResponse fromEntity(User user) {
             return UserInfoResponse.builder()
-                    .id(user.getId())
+                    .id(user.getId()) // 이제 user.getId() (Long)와 타입이 일치합니다.
                     .email(user.getEmail())
                     .nickname(user.getNickname())
                     .phone(user.getPhone())
